Move deploy.py and cron_runner.py into implementations/modal/
Move Modal-specific deployment code into the implementations/modal/
subdirectory, matching the directory structure for provider-specific
implementations. Local scheduling (via cron directly) would go in
implementations/local/ when implemented.
